+++ lib/libutils/utils/crypto/encryption_helper.h
@@ -19,8 +19,8 @@ void decrypt(EncryptedMessage& cipher,
 	     const std::string& key,
 	     common::ByteArray& plain_text);
 
-int get_cipher_length(int plain_length);
-int get_plain_length(int cipher_text_length);
+unsigned long long get_cipher_length(unsigned long long plain_length);
+unsigned long long get_plain_length(unsigned long long cipher_text_length);
 int get_key_length(void);
 int get_padding_length(void);
